Subdomainクラスをテンプレート化
テスト(モック)のために具象クラス(Hash)のメソッドを仮想化してしまっていた．
モックを独立したクラスにすることで仮想化しなくてもよくなった．
しかしSubdomainクラスをテンプレートクラスにせざるをえなくなった．
