+++ src/Utils.cpp
@@ -1,17 +1,11 @@
-#include "Subdomain.hpp"
-#include <cmath>
 #include <cstddef>
-#include <memory>
 #include <string>
-#include "Hash.hpp"
+#include "Utils.hpp"
 
 constexpr char kCharacters[] = {'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j',
                                 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't',
                                 'u', 'v', 'w', 'x', 'y', 'z'};
 constexpr std::size_t kSize = sizeof(kCharacters) / sizeof(char);
-
-namespace
-{
 
 std::string get_character(std::size_t value, std::size_t current)
 {
@@ -32,23 +26,7 @@ std::string get_subdomain(std::size_t hash, std::size_t total)
   return get_character(hash, total);
 }
 
-}  // anonymous namespace
-
-Subdomain::Subdomain(std::shared_ptr<Hash> hash, const std::string& domain, std::size_t length)
-  : hash_(hash)
-  , domain_(domain)
-  , length_(length)
+std::size_t get_size(void)
 {
-}
-
-std::string Subdomain::get(void) const
-{
-  if (length_) {
-    std::size_t total = pow(kSize, length_);
-    std::size_t hash = hash_->hash(domain_, total);
-
-    return get_subdomain(hash, total);
-  } else {
-    return "";
-  }
+  return kSize;
 }
