+++ src/Utils.hpp
@@ -1,0 +1,7 @@
+#ifndef UTILS_H
+#define UTILS_H
+
+extern std::string get_subdomain(std::size_t hash, std::size_t total);
+extern std::size_t get_size(void);
+
+#endif  // UTILS_H
